Minor changes in folder structure
